Fix banded example
And move pairwise_correlation
