+++ naplib/stats/__init__.py
@@ -1,7 +1,7 @@
-from .encoding import discriminability
+from .encoding import discriminability, pairwise_correlation
 from .mixedeffectsmodel import LinearMixedEffectsModel
 from .pvalues import stars
 from .responsive_ttest import responsive_ttest
 from .ttest import ttest
 
-__all__  = ['discriminability','LinearMixedEffectsModel','stars','responsive_ttest', 'ttest']
+__all__  = ['discriminability','pairwise_correlation','LinearMixedEffectsModel','stars','responsive_ttest', 'ttest']
